software/Device_Validator/requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
scapy-python3==0.20
six==1.10.0
